Display extension info for "create now" overlay
