+++ packages/apps/src/overlays/Accounts.tsx
@@ -7,9 +7,9 @@ import { ApiProps } from '@polkadot/ui-api/types';
 import { SubjectInfo } from '@polkadot/ui-keyring/observable/types';
 
 import React from 'react';
-import { Trans } from 'react-i18next';
 import { Link } from 'react-router-dom';
 import styled from 'styled-components';
+import ExtensionInfo from '@polkadot/app-accounts/Banner/Info';
 import { withApi, withMulti, withObservable } from '@polkadot/ui-api';
 import accountObservable from '@polkadot/ui-keyring/observable/accounts';
 
@@ -60,16 +60,16 @@ class Accounts extends React.PureComponent<Props, State> {
         className={className}
         icon='users'
       >
-        <Trans i18nKey='noAccounts'>
-          You don&apos;t have any accounts. Some features are currently hidden and will only become available once you have accounts.
-          {' '}
+        <p>{t("You don't have any accounts. Some features are currently hidden and will only become available once you have accounts.")}</p>
+        <ExtensionInfo />
+        <p>
           <Link
             to = {{ pathname: '/accounts', state: { isCreateOpen: true } }}
             onClick={this.onClose}
           >
-            {t('Create an account now.')}
+            {t('Create an account now inside this app.')}
           </Link>
-        </Trans>
+        </p>
       </BaseOverlay>
     );
   }
